fix q14

+++ set2/14/q14final.cc
@@ -14,7 +14,7 @@ int main(int argc, char *argv[])
     {
         cout << comb + 1 << ": ";   // Display the numbered line.
 
-        for (size_t idx = 0; idx != (argc - 1); ++idx)
+        for (size_t idx = 0; (int)idx != (argc - 1); ++idx)
         {                           // Use bitwise AND to check whether argv at
                                     // idx has to be displayed in the 
                                     // combination.
